requirements.txt
streamlit
langchain
pypdf2
python-dotenv
openai
tiktoken
faiss-cpu
streamlit-chat
tabulate
loguru
unstructured
pdf2image
pytesseract
chromadb